requirements.txt
google-genai
mss
pynput
Pillow
